requirements.txt
pandas==2.3.3
playwright==1.55.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
boto3==1.40.51
rollbar==1.1.0